requirements.txt
discord.py==2.7.1
python-dotenv==1.2.2
requests==2.33.1
pytz==2026.1.post1
pandas==3.0.2
pandas_market_calendars==5.3.2
